Fixed issue where total carbs were cast to int.

+++ app/src/main/java/com/example/afg1/Meal.java
@@ -53,7 +53,7 @@ public class Meal implements Serializable {
      * @return the totalCarbs
      */
     public double getTotalCarbs() {
-        int temp = 0;
+        double temp = 0;
         for (Order o: meal){
             temp+=o.getTotalCarbs();
         }
